// GestionDuStock_UI/GestionDuStockUI/src/app/services/models/purchase-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PurchaseItemDto } from './purchase-item-dto';
export interface PurchaseDto {
  approved?: boolean;
  id?: number;
  invoiceNumber?: string;
  purchaseDate?: string;
  purchaseItems: Array<PurchaseItemDto>;
  status?: string;
  supplierId?: number;
  totalAmount?: number;
}